requirements.txt
# Browser Automation with LLM
# Core dependencies

# Browser automation
playwright==1.49.0

# Google Gemini API
google-generativeai==0.8.3

# Environment management
python-dotenv==1.0.1

# Data validation
pydantic==2.10.2

# CLI
click==8.1.7

# Logging
loguru==0.7.2

# Terminal output
rich==13.9.4

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0

# Dashboard
streamlit==1.39.0

